refactor: separate tables
